[alert sub view] Fix bug where hovering over analysis link caused error
This was the result of passing an actual event as the selection into
the analysis link rather than an EventSet that contained the selection.

(All selections should be event sets.)

This is a duplicate of https://codereview.chromium.org/2143153003/, but
ported to the polymer10-migration branch.

TBR=aiolos@chromium.org
NOPRESUBMIT=true
NOTRY=true
BUG=catapult:#2469

Review-Url: https://codereview.chromium.org/2141573005