[polymer] Fix side panel tab click handlers
This code on master uses panelTypes.forEach(panelType) (... init code)
as its method of iteration. When I moved it over to using a
SidePanelRegistry, I also moved it to using a for...of loop instead of
a forEach. This changed the scoping of panelTypes and made it so that
each new iteration overwrote the same panelType variable instead of
creating a new one. As a result of this, clicking each tab would
result in the last panelType being selected.

To fix this, I bound each panelType to its corresponding click handler
by using bind() and passing it as the first argument.

NOPRESUBMIT=true
NOTRY=true
BUG=catapult:#2470

Review-Url: https://codereview.chromium.org/2142383002