[polymer] Fix lag in checking alert checkbox
The slowness was caused by the fact that, every time that a checkbox was
toggled, the "highlighted" property of each alert was changed. This
caused any alerts list observers to be notified and there were many,
many alerts list observers because the alerts list was bound to the
alerts table using two way data binding. This meant the changes needed
to be propagated all the way up.

This problem was exacerbated by a regression in the logic to find a
common revision range among the selected alerts. This regression was
present in the polymer migration branch but not master. The effect was
that many more alerts were highlighted than were actually in an
intersecting range.

To fix all of this, I fixed the range intersection logic and made
"highlighted" a computed property based on a new property,
"commonRevisionRange". This meant that all observers weren't called
when the highlights changed.

The time to check a checkbox is:

  Before change: 5.8s
  After change: .1s (still not impressive, but a lot better)

NOPRESUBMIT=true
NOTRY=true
BUG=catapult:#2496

Review-Url: https://codereview.chromium.org/2171443002