[polymer] Fix a bug where columns were being incorrectly ellipsized
This was happening because computing the binding for
item.additionColumnValues.master was resulting in
"cannot compute property 'master' of undefined" when the item
was not a header for its row group.

There's still a bug where grouped alerts are showing up as ellipsized,
which I believe should be fixed with catapult:#2503.

NOPRESUBMIT=true
NOTRY=true
BUG=catapult:#2499

Review-Url: https://codereview.chromium.org/2172473002